backend/requirements.txt
fastapi
uvicorn
google-generativeai
web3
python-multipart
requests
python-dotenv
pyjwt[crypto]
passlib
bcrypt
sqlalchemy
psycopg2-binary
anyio
exifread
ffmpeg-python
geopy
cloudinary